Update data_cleaning_and_unigram
We have stored article abstracts in an excel file. this code gives us an idea of the most crucial step of data analysis which is pre-processing and data cleaning. our data is quantitative in nature. we then extracted features, used countvectorizer, TF-IDF to put a value on our feature, and finally created a unigram analysis. Finally, we showed the unigram in a histogram to showcase the importance and weight of each feature.